fix: fix valid_file_path error and theme change
Fixed valid file path error where I was incorrectly asserting filename which was
theme name as incorrect. It is now fixed with also a new autocmd which triggers
rehighlighing of statusline on theme change.

With this the statusline should be working fine with lazyvim telescope change and appropriate lualine usage as of now.

BREAKING CHANGE: Why previous version was working with the bug is unknown but it is fixed in this commit