feat: refactor tabufline and fix lualine config
Added colors for tabufline component with a dummy theme toggle as of now.
Used color functions from bufferline to adjust for default colors.

Also Fixed use_lualine which was not changing on colorscheme change.
Made it into a function which is now causing the variable to refresh
to use latest info.

Added Colors and refactored tabufline which is now usable as a standalone component