fix(modules): fix similar filename path dissection for windows
I have applied a fix which fixes path dissection which was done for unix paths.
The API vim.api.nvim_get_buf_name returns sometimes "\\", "/" or both as path
separator so I have converted windows path to unix path if OS is of type windows
as windows paths can't contain "//" or "\" in their directory names and also
changed the display of tab names from "/../" to "/" seperator when differentiator
is parent directory itself.