feat(colors): lazyload highlights of theme same as base46 nvchad
So when opening a existing file the ui loads veryfast even with priority of
colorscheme with lazy.nvim and does'nt have the necessary highlights ready.
So following base46 plugin of nvchad example I have stored the highlights in
a cache file to retrieve on startup and rewrite on theme change.