feat: add settings cache and default tokyonight colorscheme
So As nvchad UI plugin is very fast all highlights are not set by default for
cases like lazyvim where colorscheme is not mentioned.

To circumvent that and for the case where default colorscheme is not given
tokyonight defaults are set. By this we are lazyloading colorscheme for
UI plugin successfully.