feat(colors): use previous cache and load based on changes
So as we are saving highlights in cache, we are never really utilizing them
as we don't know the previous state of settings, colorscheme, dark and light
mode we can't simply use the cache and have to compute highlights everytime.

So to avoid all that we are now storing settings hash cache which stores hash
of settings and colorscheme and background mode which we than again use to
match and verify so that we can fully utilize out highlights cache.

Also as lazyvim has the theme changer which is not permanent we should also not
store cache when we simply change themes and only store and compute cache on
startup.

With this we are now effectively reusing highlights cache and avoiding
computing cache again.

But the trade-offs of opening and closing the file and theme updates not being
captured is still there.

For now I am leaving this as it is and make a decision later on this.